requirements.txt
streamlit==1.28.0
sentence-transformers==2.2.2
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
plotly==5.15.0
openpyxl==3.1.2
pdfminer.six==20221105
wordcloud==1.9.2
transformers==4.26.1
huggingface_hub
torch==2.0.1

